Index: tools/swilena/ChangeLog from  Nicolas Burrus  <burrus_n@lrde.epita.fr>
        * meta/swilena_ntg.i: Define typemaps for python.
        * meta/swilena_image.i: Don't define set, at and ref methods
        taking point2d for python and ruby.

        * python/tests/simple1.py: Adjust consequently.
        * ruby/tests/simple1.rb: Likewise.
        * doc/examples/python/simple.py: Likewise.
        * doc/examples/ruby/simple.rb: Likewise.

    prcs2git-id: 10.26

+++ tools/swilena/doc/examples/python/simple.py
@@ -13,10 +13,10 @@ swilena_display.display_command = "xv %{image_file}"
 ima = image2d_uint(5,5)
 print ima.at(5,5).value()
 
-ima.set(2,2,uint(2))
+ima.set(2, 2, 2)
 print ima
 
-p = ima.ref(1,1)
+p = ima.ref(1, 1)
 p.value(3)
 print ima
